+++ CHANGELOG.md
@@ -6,7 +6,7 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 - [#609](https://github.com/clj-kondo/clj-kondo/issues/609): typecheck var usage, e.g. `(def x :foo) (inc :foo)` will now give a warning
 - [#1867](https://github.com/clj-kondo/clj-kondo/issues/1867): add name metadata to class usage
-- [#1875](https://github.com/clj-kondo/clj-kondo/issues/1875): add `:duplicate-field` linter for deftype and defrecord definitions.
+- [#1875](https://github.com/clj-kondo/clj-kondo/issues/1875): add `:duplicate-field-name` linter for deftype and defrecord definitions.
 
 ## 2022.11.02